ml_service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic-settings
pydantic[email]
passlib[bcrypt]
python-jose[cryptography]
python-multipart
joblib
scikit-learn
redis
celery
pandas
passlib==1.7.4
bcrypt==3.2.2
prometheus-fastapi-instrumentator
prometheus_client

pytest
pytest-asyncio
httpx
pytest-cov
pytest-mock
httpx